# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tauco Cap Badak API
  version: 1.1.0
  description: |
    Contract-first REST API for the Phase 1B public backend and the Phase 1C
    local shadow CMS. Administration operations are contractual only in C1 and
    are not registered by the runtime until their implementation gate.
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /api/v1/admin/auth/login:
    post:
      tags: [Admin authentication]
      summary: Authenticate an admin with password and mandatory second factor
      operationId: adminLogin
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/AdminAuthOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/totp/setup:
    post:
      tags: [Admin authentication]
      summary: Start mandatory TOTP enrollment
      operationId: adminSetupTotp
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      responses:
        '200':
          $ref: '#/components/responses/AdminTotpSetupOK'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/totp/enable:
    post:
      tags: [Admin authentication]
      summary: Confirm TOTP enrollment and return one-time recovery codes
      operationId: adminEnableTotp
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTotpCodeRequest'
      responses:
        '200':
          $ref: '#/components/responses/AdminRecoveryCodesOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/refresh:
    post:
      tags: [Admin authentication]
      summary: Rotate the refresh token and access session
      operationId: adminRefreshSession
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      responses:
        '200':
          $ref: '#/components/responses/AdminAuthOK'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/logout:
    post:
      tags: [Admin authentication]
      summary: Revoke the current admin session
      operationId: adminLogout
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      responses:
        '204':
          $ref: '#/components/responses/AdminNoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/me:
    get:
      tags: [Admin authentication]
      summary: Get the authenticated admin and effective permissions
      operationId: adminGetMe
      security:
        - AdminCookie: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      responses:
        '200':
          $ref: '#/components/responses/AdminUserOK'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/auth/recovery-codes/regenerate:
    post:
      tags: [Admin authentication]
      summary: Replace all unused recovery codes after TOTP confirmation
      operationId: adminRegenerateRecoveryCodes
      security:
        - AdminCookie: []
          AdminCSRF: []
      parameters:
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTotpCodeRequest'
      responses:
        '200':
          $ref: '#/components/responses/AdminRecoveryCodesOK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    InternalMetricsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: A dedicated rotatable operations token; never an admin JWT.
    AdminCookie:
      type: apiKey
      in: cookie
      name: tauco_admin_access
      description: Short-lived HttpOnly, Secure, SameSite=Strict admin access cookie.
    AdminCSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit token required with AdminCookie on unsafe methods.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Optional caller-provided trace identifier. Invalid values are replaced
        by the server.
      schema:
        $ref: '#/components/schemas/RequestID'
...
  headers:
    XRequestID:
      description: Request trace identifier.
      required: true
      schema:
        $ref: '#/components/schemas/RequestID'
...
    NoStore:
      description: Prevents storage of private or operational responses.
      required: true
      schema:
        type: string
        enum:
          - no-store
    RetryAfter:
      description: Seconds until another request may be attempted.
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 60
...
  responses:
...
    BadRequest:
      description: Malformed JSON, header, cursor, or query
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            malformedRequest:
              value:
                type: urn:tauco-cap-badak:problem:bad-request
                title: Permintaan tidak valid
                status: 400
                detail: Format permintaan tidak dapat diproses.
                instance: /api/v1/contact-messages
                code: BAD_REQUEST
                requestId: 98a70ee8a919423b883c23ea275d468a
...
    ValidationFailed:
      description: One or more semantic field validations failed
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidEmail:
              value:
                type: urn:tauco-cap-badak:problem:validation
                title: Permintaan tidak valid
                status: 422
                detail: Periksa field yang ditandai.
                instance: /api/v1/contact-messages
                code: CONTACT_VALIDATION_FAILED
                requestId: 98a70ee8a919423b883c23ea275d468a
                errors:
                  - field: email
                    code: INVALID_EMAIL
                    message: Masukkan alamat email yang valid.
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            rateLimited:
              value:
                type: urn:tauco-cap-badak:problem:rate-limit
                title: Terlalu banyak permintaan
                status: 429
                detail: Coba kembali setelah waktu yang ditentukan.
                instance: /api/v1/contact-messages
                code: RATE_LIMIT_EXCEEDED
                requestId: 98a70ee8a919423b883c23ea275d468a
    AdminAuthOK:
      description: Authenticated session; secure cookies are rotated by the server
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAuthResponse' }
    AdminTotpSetupOK:
      description: Pending TOTP enrollment material
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminTotpSetupResponse' }
    AdminRecoveryCodesOK:
      description: One-time recovery codes; the plaintext list is never returned again
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminRecoveryCodesResponse' }
    AdminUserOK:
      description: Authenticated admin identity and permissions
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminUserResponse' }
...
    AdminNoContent:
      description: Operation completed with no response body
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
...
    Conflict:
      description: Current resource state conflicts with the requested operation
      headers:
        X-Request-ID: { $ref: '#/components/headers/XRequestID' }
        Cache-Control: { $ref: '#/components/headers/NoStore' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unauthorized:
      description: Missing or invalid operations token
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
          example: Bearer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unauthorized:
              value:
                type: urn:tauco-cap-badak:problem:unauthorized
                title: Autentikasi diperlukan
                status: 401
                detail: Token operasional tidak tersedia atau tidak valid.
                instance: /internal/metrics
                code: UNAUTHORIZED
                requestId: 98a70ee8a919423b883c23ea275d468a
    Forbidden:
      description: Authenticated principal is not allowed to read metrics
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            forbidden:
              value:
                type: urn:tauco-cap-badak:problem:forbidden
                title: Akses ditolak
                status: 403
                detail: Kredensial tidak memiliki akses ke resource ini.
                instance: /internal/metrics
                code: FORBIDDEN
                requestId: 98a70ee8a919423b883c23ea275d468a
    InternalServerError:
      description: Unexpected server error with no internal detail exposed
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            internal:
              value:
                type: urn:tauco-cap-badak:problem:internal
                title: Terjadi kesalahan internal
                status: 500
                detail: Permintaan tidak dapat diproses saat ini.
                instance: /api/v1/home
                code: INTERNAL_SERVER_ERROR
                requestId: 98a70ee8a919423b883c23ea275d468a
...
  schemas:
    RequestID:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      example: 98a70ee8a919423b883c23ea275d468a
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
        - apiVersion
      properties:
        requestId:
          $ref: '#/components/schemas/RequestID'
        apiVersion:
          type: string
          enum:
            - v1
...
    ValidationError:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        code:
          type: string
          minLength: 3
          maxLength: 80
          pattern: '^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$'
        message:
          type: string
          minLength: 3
          maxLength: 240
          pattern: '^\S(?:[\s\S]*\S)?$'
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - code
        - requestId
      properties:
        type:
          type: string
          format: uri
          maxLength: 160
          pattern: '^urn:tauco-cap-badak:problem:[a-z0-9]+(?:-[a-z0-9]+)*$'
        title:
          type: string
          minLength: 3
          maxLength: 120
          pattern: '^\S(?:[\s\S]*\S)?$'
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 3
          maxLength: 500
          pattern: '^\S(?:[\s\S]*\S)?$'
        instance:
          $ref: '#/components/schemas/ProblemInstance'
        code:
          type: string
          minLength: 3
          maxLength: 80
          pattern: '^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$'
        requestId:
          $ref: '#/components/schemas/RequestID'
        errors:
          type: array
          minItems: 1
          maxItems: 32
          items:
            $ref: '#/components/schemas/ValidationError'
    ProblemInstance:
      type: string
      format: uri-reference
      minLength: 1
      maxLength: 2048
      pattern: "^/(?:$|(?:[A-Za-z0-9._~!$&'()*+,;=:@-]|%[0-9A-Fa-f]{2})(?:[A-Za-z0-9._~!$&'()*+,;=:@/-]|%[0-9A-Fa-f]{2})*)$"
      description: |
        Absolute, path-only URI-reference built from the escaped request path.
        Uppercase segments, a trailing slash, and percent-encoded octets are
        valid. Empty, non-absolute, network-path, or oversized values fall back
        to `/` before a problem response is emitted.
      x-source: escaped-request-path
...
    AdminLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      x-invariant: at-most-one-of-totpCode-or-recoveryCode
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 160
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
        totpCode:
          type: string
          pattern: '^[0-9]{6}$'
        recoveryCode:
          type: string
          pattern: '^[A-Z0-9]{4}(?:-[A-Z0-9]{4}){2}$'
    AdminTotpCodeRequest:
      type: object
      additionalProperties: false
      required: [totpCode]
      properties:
        totpCode:
          type: string
          pattern: '^[0-9]{6}$'
    AdminUser:
      type: object
      additionalProperties: false
      required: [id, email, status, mfaEnabled, roles, permissions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, disabled]
        mfaEnabled:
          type: boolean
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            pattern: '^[a-z][a-z0-9_]*$'
        permissions:
          type: array
          uniqueItems: true
          items:
            type: string
            pattern: '^[a-z][a-z0-9]*([.:_-][a-z0-9]+)*$'
    AdminAuthData:
      type: object
      additionalProperties: false
      required: [status, user, expiresAt]
      properties:
        status:
          type: string
          enum: [authenticated, mfa_setup_required]
        user:
          $ref: '#/components/schemas/AdminUser'
        expiresAt:
          type: string
          format: date-time
    AdminAuthResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminAuthData' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    AdminUserResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminUser' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    AdminTotpSetupData:
      type: object
      additionalProperties: false
      required: [otpauthUri, manualKey, expiresAt]
      properties:
        otpauthUri:
          type: string
          format: uri
          maxLength: 2048
        manualKey:
          type: string
          pattern: '^[A-Z2-7]{16,64}$'
        expiresAt:
          type: string
          format: date-time
    AdminTotpSetupResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminTotpSetupData' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    AdminRecoveryCodesData:
      type: object
      additionalProperties: false
      required: [codes]
      properties:
        codes:
          type: array
          minItems: 8
          maxItems: 12
          uniqueItems: true
          items:
            type: string
            pattern: '^[A-Z0-9]{4}(?:-[A-Z0-9]{4}){2}$'
    AdminRecoveryCodesResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AdminRecoveryCodesData' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...